fix: rename file to match DetailLabelButton component name
